move build command to its own file
