+++ .ops/main.go
@@ -16,30 +16,3 @@ func main() {
 func (Ops) Hello() {
 	println("Hello world!")
 }
-
-func (Ops) Build () {
-	var rnr = sys.Runner().WithEnv(map[string]string{
-		"PKGNAME": "cmdio",
-	})
-	defer rnr.Close()
-
-	err := rnr.Run("echo", "hello from", rnr.Env("PKGNAME"))
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	err = rnr.Run(useNpmOrPnpm(rnr), "install")
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	err = rnr.Run(useNpmOrPnpm(rnr), "run", "build")
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	err = rnr.Run("echo", "goodbye from", rnr.Env("PKGNAME"))
-	if err != nil {
-		log.Fatal(err)
-	}
-}
